fix(vm): only promote to int64/uint64 when both operands match signedness
Avoid mixed-sign regression where int64 + uint64 would incorrectly
promote to uint64, causing wrong results for negative values.

+++ expr_test.go
@@ -805,11 +805,11 @@ func TestExpr(t *testing.T) {
 		},
 		{
 			`Uint64 + 0`,
-			uint64(0),
+			0,
 		},
 		{
 			`Uint64 + Int64`,
-			uint64(0),
+			0,
 		},
 		{
 			`Int32 + Int64`,
